Rename module: 'mappings' -> 'maps'
