+++ dendroid/red_black.py
@@ -12,7 +12,7 @@ from reprit.base import generate_repr
 from .core.abcs import (NIL,
                         Node as _Node,
                         Tree as _Tree)
-from .core.mappings import map_constructor as _map_constructor
+from .core.maps import map_constructor as _map_constructor
 from .core.sets import set_constructor as _set_constructor
 from .core.utils import (dereference_maybe as _dereference_maybe,
                          maybe_weakref as _maybe_weakref,
