+++ dendroid/splay.py
@@ -8,7 +8,7 @@ from .binary import Node
 from .core.abcs import (NIL,
                         AnyNode,
                         Tree as _Tree)
-from .core.mappings import map_constructor as _map_constructor
+from .core.maps import map_constructor as _map_constructor
 from .core.sets import set_constructor as _set_constructor
 from .core.utils import (to_unique_sorted_items as _to_unique_sorted_items,
                          to_unique_sorted_values as _to_unique_sorted_values)
